+++ conSERN/nodegen.h
@@ -43,8 +43,8 @@ typedef struct
 {
     float    *x;
     float    *y;
-    uint64_t count;
-    uint64_t start;
+    uint32_t count;
+    uint32_t start;
     uint32_t i;
     uint32_t j;
     double  p;
